CleanUp Db Connection Handling
Move db and queries to its own package
